Add search bar popup for mobile
